Better selection of data types especially json

+++ demo/compound_sql.go
@@ -407,7 +407,7 @@ const sqlCreateDbCompoundTableSqlite = `
 CREATE TABLE %s%s%s (
  alpha    text,
  beta     text,
- category integer
+ category tinyint unsigned
 )
 `
